Revised Homework2
Revisions done to: 1. use body tag for the most common text content properties, 2. correct colors in the hover state of elements and also other styling details  to be in sync with the style guide page in Figma, 3.streamline portfolio page from 2 sections to just 1 and also used <ul> and <li> tags for the said section.